Revert "Merge pull request #19 from JUNCapstoneProject/LoginPage"
This reverts commit 5636e3d4ebe04351f2f6f012e009663f594d119b, reversing
changes made to 981edf9348275a83cce64a62e5456d7c8542948d.
