+++ src/App.jsx
@@ -8,8 +8,6 @@ import { useEffect } from "react";
 import { useDispatch } from "react-redux";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import RouteWatcher from "./components/common/RouterWatcher";
-import LoginPage from "./pages/LoginPage";
-import SignupPage from "./pages/SignupPage";
 
 // 레이아웃 컴포넌트
 import Navbar from "./components/layout/navbar";
@@ -84,8 +82,6 @@ function App() {
           <Route path="/report/:reportId" element={<ReportDetail />} />
           <Route path="/report/:reportId/edit" element={<ReportEdit />} />
           <Route path="/report/create" element={<ReportCreate />} />
-          <Route path="/login" element={<LoginPage />} />
-          <Route path="/signup" element={<SignupPage />} />
         </Routes>
       </div>
     </Router>
