+++ src/components/common/SignupModal.css
@@ -1,0 +1,96 @@
+.modal-overlay {
+    position: fixed;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background-color: rgba(0, 0, 0, 0.5);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    z-index: 1000;
+  }
+  
+  .modal-content {
+    background: white;
+    padding: 2rem;
+    border-radius: 8px;
+    width: 100%;
+    max-width: 400px;
+    position: relative;
+  }
+  
+  .modal-close {
+    position: absolute;
+    top: 1rem;
+    right: 1rem;
+    background: none;
+    border: none;
+    font-size: 1.5rem;
+    cursor: pointer;
+  }
+  
+  .form-group {
+    margin-bottom: 1rem;
+  }
+  
+  .form-group label {
+    display: block;
+    margin-bottom: 0.5rem;
+  }
+  
+  .form-group input {
+    width: 100%;
+    padding: 0.5rem;
+    border: 1px solid #ddd;
+    border-radius: 4px;
+  }
+  
+  .signup-submit-btn {
+    width: 100%;
+    padding: 0.75rem;
+    background: #000;
+    color: white;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+    margin-top: 1rem;
+  }
+  
+  .login-link {
+    text-align: center;
+    margin-top: 1rem;
+  }
+  
+  .password-requirements {
+    margin-top: 10px;
+    font-size: 0.9em;
+  }
+  
+  .password-requirements ul {
+    list-style: none;
+    padding-left: 0;
+  }
+  
+  .password-requirements li {
+    margin: 5px 0;
+    padding-left: 20px;
+    position: relative;
+  }
+  
+  .password-requirements li::before {
+    content: "✕";
+    position: absolute;
+    left: 0;
+    color: #ff4444;
+  }
+  
+  .password-requirements li.valid::before {
+    content: "✓";
+    color: #4CAF50;
+  }
+  
+  .signup-submit-btn:disabled {
+    background-color: #cccccc;
+    cursor: not-allowed;
+  }
